Allowing time object validation to be skipped
I'm using the blind spot of timeObject to string conversion to force the 24 hour and 12 hour times to line up when sending back the return time object.

This means that the time object I send in for conversion will technically be invalid so I need to skip the validation step.