Removing hrs24 from manual entry log
This is the manual entry log which means that the user has to be able to see the value. Generally the hrs24 value is never visible to the user. So it was pointless distinguishing between hrs24 and hrs12 when hrs12 was the only hrs that was likely to be used.